move CombinedBuilder to correct namespace
